prevent delete if has posts

+++ app/Http/Controllers/CategoriesController.php
@@ -94,6 +94,12 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->posts->count() > 0) {
+          session()->flash('error', 'Category cannot be deleted because it has some posts.');
+
+          return redirect()->back();
+        }
+        
         $category->delete();
 
         session()->flash('success', 'Category deleted successfully.');
